[MultiViews][m]: renamed DataPackageView to MultiViews, made it to render only graphs, it does not load any data now, updated tests and snapshot - refs #126
